fix: ignore errors on duplicate `SetHeader` calls
Updating gRPC library to 1.47 version makes it more picky about the
duplicate calls to `SetHeader`. Due to the nature of proxying, we might
try to set it multiple times, so simply ignore the error (doing best
effort to set it correctly, but with multiple upstreams setting headers
from all upstreams is not possible).

Signed-off-by: Andrey Smirnov <andrey.smirnov@talos-systems.com>